kinomeDb/requirements.txt
djangorestframework==3.15.2
django-cors-headers==4.4.0
